Fix error passing namespace to csv-generator
gen_csv took 3 fixed arguments, but we were only passing it 2. We were
also not using the third argument. This meant that the first entry in
operatorArgs was dropped, which was always --namespace=openstack. This
only worked because none of the csv-generators were validating their
input, and they all defaulted to openstack anyway.

Additionally the default namespace we were failing to set was
incorrectly set to "openstack-cluster-operator".

Additionally, because gen_csv is executed in a subshell the calling
shell's `set -e` had no effect. If the execution of a csv-generator
failed. This left no output for that generator with no indication of
what caused the error.