Add olm build Dockerfile's
Older versions of these files in the git checkout could cause issues
with not always using the latest versions. As it's best practice to
run "make clean" before re-running the olm scripts, just have
that target also delete any of the old olm Dockerfile's.

dev-tools does a "make clean" in the olm.yaml playbook, so this commit
should help with issues around unintentionally using old versions of the
Dockerfile's.

Signed-off-by: James Slagle <jslagle@redhat.com>